feat: add a simple dry-run flag

+++ internal/fs/fs.go
@@ -6,9 +6,9 @@ import (
 )
 
 type FileSystemAdapter interface {
-	IsValidName(name string) bool    // we need a way to tell if file name is valid
-	SanitizeName(name string) string // we need a way to sanatize name
-	IsCaseSensitive() bool           // we need a way to tell if naming IsCaseSensitive, is different per os
+	IsValidName(name string) bool
+	SanitizeName(name string) string
+	IsCaseSensitive() bool
 }
 
 type RenameOp struct {
@@ -16,12 +16,16 @@ type RenameOp struct {
 	NewPath string
 }
 
-func Apply(ops []RenameOp) error {
+func Apply(ops []RenameOp, dryRun bool) error {
 	for _, op := range ops {
-		if err := os.Rename(op.OldPath, op.NewPath); err != nil {
-			return fmt.Errorf("failed to rename %s to %s: %w", op.OldPath, op.NewPath, err)
+		if dryRun {
+			fmt.Printf("Would rename: %s -> %s\n", op.OldPath, op.NewPath)
+		} else {
+			if err := os.Rename(op.OldPath, op.NewPath); err != nil {
+				return fmt.Errorf("failed to rename %s to %s: %w", op.OldPath, op.NewPath, err)
+			}
+			//TODO: a mode to log on success?
 		}
-		fmt.Printf("Renamed: %s -> %s\n", op.OldPath, op.NewPath)
 	}
 	return nil
 }
